Share package installation directory across notebooks in the same Python environment.
This allows build intermediate files to be shared. In the typical case
when a user is restarting the same notebook multiple times, the install
time of the first cell that contains all package dependencies will be
reduced.

To create some isolation but still provide some utility, a different
package installation directory is used per Python environment. The
directory is determined by guessing the type of environment the
notebook is running in. The following scenarios have been considered so
far:

- Python 3 running inside a virtualenv. The installation directory will
be a sibling of the directory that contains the environment.
- Anaconda. In this case, the installation directory will be stored
inside the Anaconda environment directory.
- System-wide Python or unknown environment. The ~/swift-env directory
will be used.

Limitations and To-Dos:
* Two notebooks running at the same time inside the same environment
might still clobber each other.
* Ideally, we should provide a way for the user to override the
installation directory and define an arbitrary one.
* Check how to identify Python2 + virtualenv.